tested stream

+++ stream.h
@@ -2,6 +2,7 @@
 #ifndef VP_STREAM
 #define VP_STREAM
 
+#include <stdlib.h>
 #include "request.h"
 
 typedef struct {
@@ -12,7 +13,7 @@ typedef struct {
 
 stream *stream_init(int br);
 void stream_add_request(stream *s, request *r);
-void stream_calc_throughput(stream *s, int alpha);
+void stream_calc_throughput(stream *s, float alpha);
 void stream_free(stream *s);
 
 #endif
